[#196]:svarga:feature, linear_value_dataset round-trip for iterator-only STL containers
H5Mstl.hpp: forward_list<T,A> size overload via std::distance; explicit
data(array<T,N>&) / data(const array<T,N>&) overloads returning T*.

H5Tmeta.hpp: fix narrowing cast in iterators access_traits_t::size() —
std::distance returns ptrdiff_t, cast to size_t for the array initializer.

H5Dread.hpp: add access_t::iterators dispatch in read<T>(ds_t) — reads into
a flat vector<element_t> then constructs T via range constructor; forward_list
uses assign(begin,end) since it has no range constructor that matches.

test/H5Dio.cpp: replace linear_value_dataset stub with live round-trips for
vector<int> and all access_t::iterators types (list, forward_list, deque,
set, multiset, unordered_set, unordered_multiset). std::array<int,4> retains
its WARN stub — rank-0 write path bug is a separate issue.

+++ h5cpp/H5Tmeta.hpp
@@ -708,7 +708,7 @@ namespace h5::meta {
             if constexpr (meta::has_size<T>::value)
                 return {static_cast<std::size_t>(c.size())};
             else
-                return {std::distance(c.begin(), c.end())};
+                return {static_cast<std::size_t>(std::distance(c.begin(), c.end()))};
         }
     };
